requirements.txt
pyobjc-core
pyobjc-framework-SystemConfiguration